Get rid of security.py
Move these two helper functions into local helpers.
