+++ lms/views/helpers/__init__.py
@@ -1,4 +1,5 @@
+from lms.views.helpers._authentication import check_password
 from lms.views.helpers._canvas_files import canvas_files_available
 
 
-__all__ = ("canvas_files_available",)
+__all__ = ["check_password", "canvas_files_available"]
